refactor(ui): вынесена shell-level navigation policy из RootView
compact navigation и detail destination resolution перенесены в ReadingShellNavigationState, а RootView очищен от вспомогательной shell-логики
shell и sidebar view files перегруппированы по feature-папкам, а тесты обновлены под новые screen-shell сущности без изменения поведения навигации
